fix: npm storybook script runs successfully
Had to move the storybook configuration and stories dir into the same
dir of the package.json file containing the storybook script.

Also added storybook addons as dev dependencies for the whole thing to
work and a couple of story file edits.

Going to refactor the directory structure of storybook files later.